- Fixed std::sqrtf and std:atan2f. These are now std::sqrt and std::atan2 (they should still call the single precision version).
Former-commit-id: 51df517be9f4a3ede0a216e9299ce7f365d8b177

+++ trunk/Box2D/Box2D/Common/b2Math.h
@@ -57,8 +57,8 @@ inline float32 b2InvSqrt(float32 x)
 	return x;
 }
 
-#define	b2Sqrt(x)	std::sqrtf(x)
-#define	b2Atan2(y, x)	std::atan2f(y, x)
+#define	b2Sqrt(x)	std::sqrt(x)
+#define	b2Atan2(y, x)	std::atan2(y, x)
 
 inline float32 b2Abs(float32 a)
 {
